feat(cartList):삭제 시 confirm창으로 다시 확인

+++ src/components/CartPage/CartItem/index.tsx
@@ -24,8 +24,12 @@ const CartItem = ({ product, id, quantity: initialQuantity }: CartItemType) => {
   }, [quantity]);
 
   const removeItemFromCart = () => {
-    deleteInCart(id);
-    deleteInCheckedCart(id);
+    const yes = confirm(`${name}을(를) 장바구니에서 비우시겠습니까?`);
+
+    if (yes) {
+      deleteInCart(id);
+      deleteInCheckedCart(id);
+    }
   };
 
   return (
